requirements.txt
django==2.0
djangorestframework==3.11
requests